# src/main/kotlin/com/fsg/cacheservice/api/dto/ErrorResponseDto.kt
package com.fsg.cacheservice.api.dto

import java.time.Instant

data class ErrorResponseDto(
    val error: ErrorDetailsDto
)

data class ErrorDetailsDto(
    val code: String,
    val message: String,
    val timestamp: Instant,
    val path: String
)